+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed config error
+- Updated to 1.21.1